ExprDiff

Description

Section : Diff Expr

class DiffExpr a where

This class operates over the Expr data type

Minimal complete definition

eval, simplify, partDiff

Methods

eval :: Map String a -> Expr a -> a

Evaluate an expression fully, values for all variables of the expression must be supplied otherwise or an error will be thrown use simplify if you cannot supply all values

simplify :: Ord a => Map String a -> Expr a -> Expr a

Simplifies an expressions replacing variables the values in the given map, and partially evaluating when possible

usimplify :: Ord a => Expr a -> Expr a

Unary simplification, uses an empty map.

Equivalent to

simplify (Map.fromList [])

partDiff :: String -> Expr a -> Expr a

performs partial differentiation with respect to the given String identifier for a variable

(!+) :: Expr a -> Expr a -> Expr a

Equivalent to Add constructor

(!*) :: Expr a -> Expr a -> Expr a

Equivalent to Mult constructor

exCos :: Expr a -> Expr a

Equivalent to Cos constructor

exSin :: Expr a -> Expr a

Equivalent to Sin constructor

exNExp :: Expr a -> Expr a

Equivalent to NExp constructor

exLn :: Expr a -> Expr a

Equivalent to Ln constructor

(!^) :: Expr a -> Expr a -> Expr a

Equivalent to Exp constructor

val :: a -> Expr a

Equivalent to Const constructor

var :: String -> Expr a

Equivalent to Var constructor

Shoehorning stuff to get DiffExpr to work with integers

class Num a => ShoeHornFloating a where

A class to shoehorn the integral types into the floating type (for some function)

Minimal complete definition

shoeHornCos, shoeHornSin, shoeHornLn, shoeHornNExp, shoeHornExp

Methods

shoeHornCos :: a -> a

Cosine function

shoeHornSin :: a -> a

Sine function

shoeHornLn :: a -> a

Natural Logarithm function, in haskell log

shoeHornNExp :: a -> a

Natural Exponentiation functoion, in haskell exp

shoeHornExp :: a -> a -> a

Exponentiation function, in Haskell '(**)'