ExprNorm

Description

Constains some helper functions used by simplify in ExprDiff.

Used to normalize expressions to a constant form that can be further reduced by simplify easily

Documentation

Multiplication Normalization

multNorm

Arguments

:: (Ord a, Num a)

=> Expr a
An Expression of form (Mult e1 e2) any other form will return itself

-> Expr a
A normalized form of (Mult e1 e2) or a Constant if it is a multiplication of constants else the input

Normalize a multiplication Expression Always converts a Mult Expression to another Mult Expression or Const Expression in normalized form. (Note this isn't a proper normalized form in terms of rewrite functions but normalizes it to a form that simplify can normalize) This is a recursive normalization function, there is an alternate multiplication normalization function that uses lists.

expandExp :: (Num a, Ord a) => Expr a -> Expr a

Expands an exponent with multiplication inside to have multiplication outside

fromListMult :: [Expr a] -> Expr a

Converts a list of expressions into a multiplication expression, fails if given an empty list

toListMult :: Expr a -> [Expr a]

Converts a multiplication Expression Mult into a list of its components

multNorml

Arguments

:: (Ord a, Num a)

=> [Expr a]
list of expressions multiplied together, use toListMult to convert from an Expr a

-> [Expr a]
list of expressions multiplied together, use fromListMult to convert to an Expr a

Normalize a multiplication Expression Always converts a Mult Expression to another Mult Expression or Const Expression in normalized form. This is a normalization function using lists, use this to understand tha basics of how this multiplication normalization works multNorm is basically this but w/o lists

Exponent Normalization

expNorm

Arguments

:: (Num a, Ord a)

=> Expr a
Exp a b expression, any other types are ignored

-> Expr a
Normalized form of Exp expression

Normalize an exponent. Always converts an exponent to another exponent

Addition Normalization

addNorm :: (Ord a, Num a) => Expr a -> Expr a

Normalize an addition Expression Always converts a Add Expression to antoher Add expression in normalized form. See addNorml

fromListAdd :: [Expr a] -> Expr a

Converts a list of expressions into an addition expression, fails if given an empty list

toListAdd :: Expr a -> [Expr a]

Converts an addition Expression Add into a list of its components

addNorml

Arguments

:: (Ord a, Num a)

=> [Expr a]
list of expressions that will be added together, use toListAdd to convert from Expr a

-> [Expr a]
list of expressions that will be added together, use fromListAdd to convert to Expr a

Normalize an addition Expression Always converts an Add Expression to another Add Expression or Const Expression in normalized form. (Note this isn't a proper normalized form in terms of rewrite functions but normalizes it to a form that simplify can normalize) Uses lists as that's the easiest way to normalize for addition.

almostEqual :: (Ord a, Num a) => Expr a -> Expr a -> Bool

A function whose name gives you no idea what the heck it's for (you always have to have nonsensical names in your code)

mCoef

Arguments

:: (Ord a, Num a)

=> Expr a
Expression to extract coefficient from

-> a
Coefficient

Extract coefficient from multipication term

For a multiplication expression this is either the constant value or 1. For any other expressions this is 1

i.e. coef of \[5x\] is 5

>>> mCoef (Mult (Const 5) (Var "x"))
5

mTerm

Arguments

:: (Ord a, Num a)

=> Expr a
Expression to extrac the term from

-> Expr a
Extracted term

Extract Term from a multiplication term

For a mutliplication expression this is everything except the constants. FOr any other expression this is itself

i.e. term of \[5x\] is x

>>> mTerm (Mult (Const 5) (Var "x"))
(var "x")

expandMult

Arguments

:: Eq a

=> [Expr a]
list of expressions that will be multiplied together, use toListMult to convert from an Expr a

-> [Expr a]
list of expressions that will be added together, use fromListAdd to convert to an Expr a form

Takes a list of expressions multiplied and returns a list of addition expanding any Add within the list of multiplication

Ln Normalization

lnNorm :: Expr a -> Expr a

Normalizes a Ln expresssion by expanding Multiplication within and bringing down the exponent for exponents within

Orphan instances

(Ord a, Num a) => Ord (Expr a)
Basic ordering, wherein things are sorted from least to greatest from most specific to least Const < Var < Cos < Sin < Ln < NExp < Add < Mult < Exp

Methods

compare :: Expr a -> Expr a -> Ordering

(<) :: Expr a -> Expr a -> Bool

(<=) :: Expr a -> Expr a -> Bool

(>) :: Expr a -> Expr a -> Bool

(>=) :: Expr a -> Expr a -> Bool

max :: Expr a -> Expr a -> Expr a

min :: Expr a -> Expr a -> Expr a