ExprParser

Description

This module contains everything used to parse a string into an expression. For explanation of how the expression is shown in the REPL output see ExprPretty

Basic binary operators

!+ : Creates an Add expression and is left associative

to parse [2+3+4]

>>> parseExprI "2!+3!+4"
(((val 2)!+(val 3))!+(val 4))

!* : Creates a Mult expression and is left associative

to parse \[(2) (3) (4)\]

>>> parseExprI "2!*3!*4"
(((val 2)!*(val 3))!*(val 4))

!^ : Creates a Exp expression and is left associative

to parse \[ 2^{3^4}\]

>>> parseExprI "2!^3!^4"
(((val 2)!^(val 3))!^(val 4))

With the operators exponentiation takes precedence over multiplication which takes precedence over addition

to parse \[ 1 + 5x^2 \]

>>> parseExprI "1!+5!*x!^2"
((val 1)!+((val 5)!*((Var "x")!^(val 2))))

Surrounding an expression with (...) will ensure that the stuff inside the brackets is evaluated first followed by whatever is outside

to parse \[(1+5x)^2\]

>>> parseExprI "(1!+5!*x)!^2"
(((val 1)!+((val 5)!*(Var "x")))!^(val 2))

Unary operators

Unary operators are generally of the form name(...). Parenthesis are not needed if the value inside is a single variable or constant but recommended for clarity

cos : parses a cosine expression

to parse \[cos (x) \]

>>> parseExprI "cosx"
(exCos(Var "x"))
>>> parseExprI "cos(x)"
(exCos(Var "x"))

sin : parses a sine expression

to parse \[ sin (x) \]

>>> parseExprI "sinx"
(exSin(Var "x"))
>>> parseExprI "sin(x)"
(exSin(Var "x"))

ln : parses a natural logarithm

to parse \[ ln (x)\]

>>> parseExprI "lnx"
(exLn(Var "x"))
>>> parseExprI "ln(x)"
(exLn(Var "x"))

e^ : parses a natural exponent

to parse \[ e^{(2+x)} \]

>>> parseExprI "e^(2!+x)"
(exNExp((val 2)!+(Var "x")))

Parsing other types

The Expr type does not contain division however it can be expressed by multiplying with an exponent to -1.

to parse \[ \frac{x}{y}\]

>>> parseExprI "x!*y!^-1"
((Var "x")!*((Var "y")!^(val -1)))

As the Expr type does not contain subtraction however this can be done by adding with a multiplication of -1

to parse \[x-y\]

>>> parseExprI "x!+(-1!*y)"
((Var "x")!+((val -1)!*(Var "y")))

To parse a log base you can use the log base formula of ln(a)/ln(b)

So to parse \[\log_{2}(x)\]

>>> parseExprI "ln(x)!*(ln(2)!^-1)"
((exLn(Var "x"))!*((exLn(val 2))!^(val -1)))

Constants are parsed based on the format of the individual parse see parseExprD parseExprI parseExprF parseExprInt

Variables are any alphatnumeric sequence that cannot be parsed into anything else

Documentation

parseExprD

Arguments

:: String
The string to be parsed

-> Expr Double
The resulting parsed Double Expr

Parses an expression with double constants. With the double being of the format (-)e(-)#

where (-) indicates that you can put a "-" sign in that position # indicates any number.

The decmial and the exponentiation can be omitted i.e.

4
4.12
4e12

are all acceptable

parseExprF

Arguments

:: String
The string to be parsed

-> Expr Float
The resulting parsed Float Expr

Parses an expression with float constants. With the float being of the format (-)e(-)#

where (-) indicates that you can put a "-" sign in that position, # indicates any number.

The decmial and the exponentiation can be omitted i.e.

4
4.12
4e12

are all acceptable

Note there may be loss of precision with very large or small numbers, using parseExprD may help slightly

parseExprI

Arguments

:: String
The string to be parsed

-> Expr Integer
The resulting parsed Integer Expr

Parses an expression with Integer constants. With the integer being of the format (-)#

where (-) indicates that you can put a "-" sign in that position, # indicates any number

i.e.

4
-42

are all acceptable

parseExprInt

Arguments

:: String
The string to be parsed

-> Expr Int
The resulting parsed Int Expr

Parses an expression with Int constants. With the integer being of the format (-)#

where (-) indicates that you can put a "-" sign in that position, # indicates any number

i.e.

4
-42

are all acceptable