ExprPretty

Description

This class automatically shows expressions in a maybe slightly more readable format. The string shown fully respresents the expression in the DSL established in ExprDiff.

Conversions from code format to Show format

a,b etc represent the show for expressions within

Add a b  -> ((a)!+(b))

Mult a b -> ((a)!*(b))

Exp a b  -> ((a)!^(b))

Cos a    -> (exCos(a))

Sin a    -> (exSin(a))

NExp a   -> (e^(a))

Ln a     -> (exLn(a))

Const a  -> (val a)

Var a    -> (var "a")

Orphan instances

Show a => Show (Expr a)

Methods

showsPrec :: Int -> Expr a -> ShowS

show :: Expr a -> String

showList :: [Expr a] -> ShowS