ExprTest

Description

Note not a lot of testing has been done due to running out of time, most of the basic cases should be covered, and the code here contains some things to test things further.

Documentation

sampleExpr :: Expr Double

Random sample expression \[x + y\]

sampleAddNorm1 :: Bool

Sample tests (To make sure any changes don't break basic cases)

Tests that \[x + x\] normalizes to \[2x\]

sampleAddNorm2 :: Bool

Tests that \[(x+y)(x+z)\] normalizes to \[x^2 + xy + xz + yz\]

sampleAddNorm3 :: Bool

Tests that \[2xy + 2x(x+y)\] normalizes to \[5x^2+7xy\]

sampleAddNorm4 :: Bool

Tests that \[5ln(xy)+5+ln(x)\] normalizes to \[5 + 6ln(x) + 5ln(y)\]

sampleAddNorm5 :: Bool

Tests that \[98x(5+i+ln(y)+z)+xln(xy)\] normalizes to \[490x + 98ix + 98xz + xln(x) + 99xln(y)\]

sampleMultNorm1 :: Bool

Tests that \[5 * x * cos (5) * x^2 \] normalizes to \[5x^3cos(5)\]

sampleMultNorm2 :: Bool

Tests that \[25xy*cos(x)*(x^{2^{2^{2^{2^2}}}})*(x^32)\] normalizes to \[25x^{65}y cos(x)\]

sampleMultNorm3 :: Bool

Tests that \[ln(x^2)*4*x^3*x^4\] normalizes to \[8 x^7 ln(x)\]

addNormGauntlet :: Bool

Tests all the SampleAddNorms

multNormGauntlet :: Bool

Tests all the SampleMultNorms

Various Stuff to help test things

sampleAdd :: (Show a, Ord a, Num a) => Expr a -> Expr a -> Bool

Basic testing of a sample add expression tests all possible branches

sampleMult :: (Show a, Ord a, Num a) => Expr a -> Expr a -> Bool

Basic testing of a sample mult expression tests all possible branches of both list and normal form

genBranches :: Eq a => (Expr a -> Expr a -> Expr a) -> [Expr a] -> [Expr a]

Generates all possible ways to distribute a list of either a Mult or Add expression (turns out there's a lot of different ways) Approximate complexity formula ~= 2*(n!)

verifyAddNormality :: (Show a, Ord a, Num a) => Expr a -> Bool

Verifies that an addition expression will normalize the same way no matter how the binary Add tree may be distributed

verifyMultNormality :: (Show a, Ord a, Num a) => Expr a -> Bool

Verifies that a multiplication expression will normalize the same way no matter how the binary Mult tree may be distributed

verifyMultNormlality :: (Show a, Ord a, Num a) => Expr a -> Bool

verifies that a multiplication expression will normalize the same way no matter how the binary Mult tree may be distributed

verifyNormality :: (Show a, Ord a, Num a) => (Expr a -> Expr a) -> (Expr a -> Expr a -> Expr a) -> (Expr a -> [Expr a]) -> Expr a -> Bool

Generic verify that some expression with a binary tree structure will normalize to the same thing

vnHelper :: (Show a, Eq a) => [Expr a] -> (Expr a -> Expr a) -> Expr a -> Bool

Helper for the verify normality functions, spits a verbose error detailing edge case if it finds something wrong

mathExamQuestion :: Expr Double

A question on the 1ZB3 Exam, copied after having to take derivative multiple times in order to verify it (gets very messy)

verifyExamQuestion :: Bool

Question = d^2g(u,v)/dudv where g (u,v) = (u+v) / (uv^2) where u = 1, v = 3, answer found = 1/9 (was C or D on version 4) (forgot if it was -1/9 or 1/9, w/e was on the Exam, there was only 1 Answer w/ 1/9 as absolute value) (Assuming 1/9 cause that's what this program got) (If I got the values of u and v confused still doesn't matter for some reason swapping the two gets 1/9 as well)

simplifyProp :: Expr Double -> Bool

Simplify property, once simplified you cannot simplify any further, Causes frequent stack overflows.

isNan :: RealFloat a => Expr a -> Bool

Check if a constant in an expression is NaN, entire function will likely eval to NaN

Orphan instances

Arbitrary (Expr Double)
generic arbitrary double

Methods

arbitrary :: Gen (Expr Double)

shrink :: Expr Double -> [Expr Double]

Arbitrary (Expr Float)
generic arbitrary float

Methods

arbitrary :: Gen (Expr Float)

shrink :: Expr Float -> [Expr Float]

Arbitrary (Expr Int)
generic arbitrary int

Methods

arbitrary :: Gen (Expr Int)

shrink :: Expr Int -> [Expr Int]

Arbitrary (Expr Integer)
generic arbitrary integer

Methods

arbitrary :: Gen (Expr Integer)

shrink :: Expr Integer -> [Expr Integer]