ExprUtil

Description

Documentation

splits :: [a] -> [([a], [a])]

a function to split lists modified slightly from https://stackoverflow.com/a/22594891

flatMap :: (t -> [a]) -> [t] -> [a]

a function to flat map stuff taken from https://stackoverflow.com/questions/2986787/haskell-flatmap